refactor: Theme에서 typographies 분리
