+++ src/styles/Theme.tsx
@@ -1,0 +1,20 @@
+import { css } from 'styled-components';
+
+export const Theme = {
+  colors: {
+    HOVER_BTN: '#FF5172',
+    MAIN_PINK: '#F48096',
+    DISABLE_BTN: '#939393',
+    WHITE_TXT: '#F9F3F3',
+    MODAL_BG: '#3B3B3B',
+    CARD_BG: '#2B2B2B',
+    INPUT_BG: '#242424',
+    PAGE_BG: '#1C1B1A',
+  },
+
+  devices: {
+    DESKTOP: `(min-width: 1200px)`,
+    TABLET: `(min-width: 600px)`,
+    MOBILE: `(min-width: 390px)`,
+  },
+};
